requirements.txt
fastapi
uvicorn
requests
python-multipart
beautifulsoup4
gspread
oauth2client
python-dotenv
